+++ docs/versionhistory.rst
@@ -7,6 +7,8 @@ This library adheres to `Semantic Versioning 2.0 <https://semver.org/#semantic-v
 
 - Added the ``suppress_type_checks()`` context manager function for temporarily
   disabling type checks
+- Added support for ``typing.Never`` (and ``typing_extensions.Never``)
+- Added support for ``Never`` and ``NoReturn`` in argument annotations
 - Made it possible to apply ``@typechecked`` on top of ``@classmethod`` /
   ``@staticmethod`` (PR by jacobpbrugh)
 - Changed ``check_type()`` to return the passed value, so it can be used (to an extent)